make function calls support langchain return types
