+++ CopilotKit/.vscode/cspell.json
@@ -1,5 +1,5 @@
 {
   "version": "0.2",
   "language": "en",
-  "words": ["openai", "copilotkit", "langchain"]
+  "words": ["openai", "copilotkit", "langchain", "kwargs"]
 }
